max_concurrency=5 default under ASAN; PatternBuffer.i u16->u32 (Windows MAX_PATH_BYTES=98302)
options_types/Context.rs: max_concurrency default 20->5 when
build_options::ENABLE_ASAN. Under ASAN every spawned bun child is several-x
heavier in RSS and ~2x slower to start, so describe.concurrent test files
that spawn one child per test (process-stdio-invalid-utf16, multi-run, etc.)
hit 20 live children at once and OOM the CI box. --max-concurrency still
overrides explicitly. Reverts the test-side multi-run.test.ts isASAN
workaround (23de6966725f) - the src/ default covers it and every other
concurrent-spawn test.

bake/bake_body.rs: PatternBuffer.i was u16 (comment claimed "covers all
platforms"), but on Windows MAX_PATH_BYTES = 32767*3+1 = 98302 > u16::MAX.
The const initializer size_of::<PathBuffer>() as u16 truncates to 32766, so
slice() returns ~64KiB of trailing zero bytes -> production.test.ts "path
must be without null bytes. Received C:\\...\\dist\\NULNUL...". Widen to u32.

Verified: stdio-utf16 24/0 in pid-ns (was 0/24); multi-run via default;
Windows cargo check 0 err/warn.

+++ src/runtime/bake/bake_body.rs
@@ -1578,19 +1578,21 @@ pub fn client_virtual_source() -> logger::Source {
 pub struct PatternBuffer {
     pub bytes: PathBuffer,
     // Zig: std.math.IntFittingRange(0, @sizeOf(bun.PathBuffer)) — smallest int
-    // fitting MAX_PATH_BYTES. u16 covers all platforms (max ~32768).
-    pub i: u16,
+    // fitting MAX_PATH_BYTES. On Windows MAX_PATH_BYTES = 32767*3+1 = 98302
+    // (> u16::MAX), so u32 is required; u16 would truncate the initial index
+    // to 32766 and `slice()` would return ~64 KiB of trailing zero bytes.
+    pub i: u32,
 }
 
 impl PatternBuffer {
     pub const EMPTY: PatternBuffer = PatternBuffer {
         bytes: PathBuffer::ZEROED, // TODO(port): Zig used `undefined`; uninit not const-safe
-        i: core::mem::size_of::<PathBuffer>() as u16,
+        i: core::mem::size_of::<PathBuffer>() as u32,
     };
 
     pub fn prepend(&mut self, chunk: &[u8]) {
         debug_assert!(self.i as usize >= chunk.len());
-        self.i -= u16::try_from(chunk.len()).expect("int cast");
+        self.i -= u32::try_from(chunk.len()).expect("int cast");
         self.slice_mut()[..chunk.len()].copy_from_slice(chunk);
     }
